+++ app/helpers/content_blocks_helper.rb
@@ -8,14 +8,4 @@ module ContentBlocksHelper
     end
     options
   end
-
-  def render_custom_block(key)
-    block = SiteCustomization::ContentBlock.custom_block_for(key, I18n.locale)
-    if current_user.administrator?
-      edit_link = link_to t("admin.actions.edit"), edit_admin_site_customization_content_block_path(block, return_to: request.path )
-    end
-    res = block&.body || ""
-    res << edit_link ? "<br>#{edit_link}" : ""
-    raw res
-  end
 end
